Upgrade sane to 4.0.3 (#8048)
A module called `merge` that is dependent on by transitive dependencies of `jest-haste-map` contains a vulnerability. With this PR, I am updating `sane` (the direct dependency) that will remove this vulnerability.

Note that it is still part of the overall `yarn.lock` because of react-native depending on Metro depending on an older version of `jest-haste-map`. This will eventually be all upgraded in the future, but we gotta start somewhere :)