update _filterTestPathsWithStats condition
We only want to _filterTestPathsWithStats if testPathPattern was
explicitly passed as an argument. The issue is that it's not possible to
simply null check globalConfig.testPathPattern to determine if it was
explicity passed becase that field gets populated with values passed to
findRelatedTests also. This approach compares the values in
globalConfig.testPathPattern to the values in globalConfig.nonFlagArgs
and if it is a 1 to 1 match it is assumed that testPathPattern was not
explicity passed as an argument.