Fix `toMatch` when re-using RegExp with 'g' flag
Fixes #9283

When testing a regex with the `g` flag, the regular expression is not
reset between assertions, such that subsequent assertions that should
pass, fail. The test below passes on Jest 23 but fails on 24:

```ts
it('matches consistently when re-using a regex with the "g" flag', () => {
  const regex = /foo/g;
  jestExpect('foo').toMatch(regex); // passes
  jestExpect('foo').toMatch(regex); // fails!
});
```

It appears this behavior was introduced in c3a0167. Previously, Jest
cloned the regex with `new RegExp()`, which sidestepped the issue by
using a new regex for every assertion. This restores that behavior. The
issue here is due to somewhat confusing behavior in the way that regular
expressions with the `g` or `y` flag behave. Internally, the regex
[keeps a `lastIndex`][1] and the next time it is matched against, it
will search from that index, or cycle back to 0 if no match is found:

```
> regex = /foo/g
/foo/g
> regex.test('foo')
true
> regex.lastIndex
3
> regex.test('foo')
false
> regex.lastIndex
0
> regex.test('foo')
true
> regex.lastIndex
3
```

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp/lastIndex