Document testTimeout configuration option
The `testTimeout` configuration option was documented
in the CLI docs but not the configuration docs so I wasn't
completely sure if it was available to use in my
package.json configuration, so I needed to try it and also
dig into the jest source code to confirm. This simply
documents the option in the config docs as well.

Fixes #9448