Fix routes on Windows
I noticed that on Windows, if you had one bad route in your jest.config.js, it would return 0 route found.
It looked like a bug to me, because this was not what happened on Linux, and there was no error message directly saying a route was wrong.
This simple fix ensure that all routes are checked and registered in the result array before returning anything to callback. What was happening is that the calls being asynchronous and the error check being at the beginning, when it encountered a bad route, it would return the callback and ignore the other route, wich could be ok.