feat(circus): expose expect.deadline()
c.f. #10895, #10905

Expose the test's timeout back to the the test, so it can make
decisions based on this, if it wants. Some proposed uses are
included in the above issue, but the implementation has got
rather big, and none of the rest absolutely have to be core.

This part does.