fix(jest-leak-detector): gc() even harder
jest-leak-detector declares a leak in some of our large projects,
a leak which does not exist. I have bisected it down to there
being too deep require() trees, and/or too many files loaded.

It is definitely not a timing issue; replacing the extra gc() with
a multi-second sleep does not fix it. Calling gc() from the app code
surprisingly does fix it, at least, for us. My working theory is that
the app's gc() takes out some of the actual garbage, leaving
leak-detector's gc() time to actually collect all of the code,
resulting, eventually, in no leak.

I realise this is hot garbage. I'm absolutely not going to try and
write a test for this, and I understand if you close the PR with
nelson.gif, and I will continue running with hacks, or try and
raise a PR which allows the leak-detector implementation to be
configurable.