Updating to include known issue with CI pipelines
The issue of detecting open handles on a filesystem where you do not have the correct access to view it causes a `Segmentation fault (core dumped)` error and unless this is documented, it is extremely hard to diagnose and fix. 
It's a small line, but would have saved us days in bug fixing without a correct way to solve it. 
We only came across that this was the issue by chance, as it works in every other environment set up