chore(jest-runner): Add info regarding timers to forceExited message
When a test finishes while a Node.js timer is still active, the worker
process will have to be force-killed by jest because the active timer
will keep the event loop alive.

It is often not immediately clear that such timers are the issue for this
message and the suggested method of `--detectOpenHandles` will never
detect them. Add a helpful message to point the user to this potential
issue.

Ref: https://nodejs.org/api/timers.html#timeoutunref