chore: add ts.format config
Since prettier is used as the default formatter,
these setting only affects "organize import" command.
Which is what we want.

It doesn't work completely, as this:

```ts
import type {
  JestHookSubscriber,
  UpdateConfigCallback,
  UsageData,
  WatchPlugin,
} from './types';
```

will be formatted as:

```ts
import type {
  JestHookSubscriber,
  UpdateConfigCallback,
  UsageData,
  WatchPlugin   // <-- comma removed
} from './types';
```

Running format doc (prettier) will fix that.
It is better than not having it.