Use transpileOnly for ts-node register
This will skip running TypeScript type checks when loading config files,
and will just remove the TypeScript annotations so the code can be read
by node directly. This makes Jest startup much more efficient.

Documentation:

https://typestrong.org/ts-node/docs/options/#transpileonly

This could be made even faster by adding the following config, which
will use the much faster swc compiler to to the transpilation:

```
transpiler: 'ts-node/transpilers/swc-experimental'
```

However, that will require additional dependencies, so it should
probably be done on an opt-in basis via a new configuration field, and I
do not want to take that on at this time.

Fixes #11337