Not the good promise state ?
I might be wrong, but at line 62,
"a <b>fulfilled</b> promise would not fail the test.", instead of "fulfilled" shouldn't it be "rejected".
Because we are talking about a promise that will fail the test. If the promise is fulfilled, it means our code run smoothly, so our test must be right ?

I'm still a learner, I might interpret it wrongly.