fix(jest-transformer): Ignoring writing cache errors on windows doesn't work
The ScriptTransformer has already a check to ignore errors under windows on writing cache file. But in case that the file cannot be access because to process try to write/access the same file also fs.existsSync() will return false under windows! So, we should simply remove this check. This should not make the behavior worse on windows.