fix(jest-core): always use workers in watch mode
The comment here already said one good reason to do this; another is
that if the test hard-crashes (e.g. an async error after it completes)
then using workers allows us to still watch for changes (perhaps to fix
that crash). So now we just always use workers in watch mode; this
probably worsens startup time slightly but for watch mode that's
hopefully not as much of a problem.

Fixes #13996.