fix(jest-core): don't use workers in watch mode if runInBand specified
In #14059, I added code to say that if you use watch mode, even if you
only run one test and one worker, we should still not run in band,
because then a hard-crash in the test will crash the entire watcher. But
the way the logic is written, this overrides even an explicit
`--runInBand`, which can be useful if you really want to, say, attach a
debugger to watch-mode; it's not the smoothest experience but it's
better than nothing.

In this commit I make it so that if you explicitly say `--runInBand`
that overrides everything else. (But if you just say `--maxWorkers=1`,
or you just only have one test to run, we still use workers.) This
required some replumbing; let me know if there's a better way.