Abort `sl root` call if output resembles a steam locomotive
Jest detects whether a repository is a sapling repo by calling the `sl`
binary, and getting the output. If `sl` (steam locomotive) is installed,
the output of `sl root` 1) takes forever to get and 2) is not the root,
but a moving image of a steam locomotive. This change monitors the
stdout stream, and aborts the `sl` call if the first character is an
escape character, which indicates that the terminal is being cleared to
make way for a train to come through.

See also: https://github.com/jestjs/jest/issues/14046