fix: allow Node16/NodeNext/Bundler moduleResolution in project's tsconfig (#14739)
This small change fixes an issue that makes using jest impossible
with inside a TypeScript project which is working on 
NodeNext, Node16 or Bundler moduleResolution setting. 

The default behaviour of ts-node is to read the default
tsconfig.json file from the project.
With a hardcoded option of "module: CommonJs" the TypeScript compiler
will throw an error, because the modern moduleResolution options
used in the project are not compatible with
the hardcoded module value.

The only way to use jest in such a repo is to change the jest.config.ts file into a JS one
(or pass the options in a different way), so that the
code responsible of instantiating ts-node is not invoked.

This commit fixes the issue by providing a missing complementary option,
moduleResolution: Node, which will work perfectly with module: CommonJs
and will not be overridden by any project-specific value in tsconfig.json.