Update MockFunctions.md
I struggled with the intro docs mentioning the word "snapshot". Hasn't been mentioned before in the intro and I don't know what it means. Feel removing this is best for minimizing confusion/friction. If knowledge of a snapshot is needed in the introduction best to make a small write up on it in the intro docs before the word 'snapshot' is mentioned