revert & fix: correct a broken link about Express
The example for Express.js was removed in the following pull request due to a broken link:
https://github.com/jestjs/jest/pull/15270

Although the domain is different,
I found what appears to be the same blog based on the author’s name and title.
Therefore, I reverted the Express.js section and corrected the link.