Update resolver docs
The `packageFilter` option to `defaultResolver` appears to have gone away as part of #15619. The versioned 30.0 docs were correctly updated, but the "Next" docs were not.

Shorten the URL comment to make it less likely to induce scrolling.

I also added another example of a resolver that I've found useful in my own project. Feedback/corrections welcome.