fix: use parsed source text for Prettier formatting
In Prettier's `parse` function, the input source text is preprocessed using
`normalizeInputAndOptions`. This normalizes line endings and therefore affects
the AST positions if CLRF was replaced with LF, typically in Windows environments.
Consequently, formatting the parsed AST has to use the preprocessed source text
in order for the AST positions to align, as otherwise the formatted output is
likely to be mangled.