Support jsdom v27 and upgrade dependency
This change upgrades the minimum required version of jsdom to
v27 (previously, v26.1).

The new version of jsdom changed the API of VirtualConsole such that the
sendTo function was renamed to forwardTo and its arguments altered.

Details about the API change are available in the pull request at:
https://github.com/jsdom/jsdom/pull/3866

Details on the latest version are available here:
https://github.com/jsdom/jsdom/releases/tag/27.0.0

The new version of jsdom dropped support for node 18, so the same is
being done for this project to match.